requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.3
SQLAlchemy==2.0.32
python-barcode==0.15.1
Pillow==10.4.0
